Add branch/commit URLs to 'but log' output
Show branch and commit web links in the but log output when a remote is available, including Gerrit-aware URLs. This adds URL generation helpers (crates/but/src/url_utils.rs), wires remote/gerrit detection into the log output, and prints per-branch and per-commit URLs (GitHub/GitLab/Gerrit) in the commit graph. It also reuses the new url_utils in push output and cleans up removed duplicate URL helper code from push/mod.rs.
